refactor(pipeline/network): extract EndpointState from NetworkDiscovery
Move endpoint URL builders and mutable closure-cell accessors out of
NetworkDiscovery.ts into a dedicated EndpointState/ sub-module:

  • FULL_TXN_PARAMS / findTxnUrlWithAccountId /
    buildTxnUrlFromTraffic / buildBalUrlFromTraffic — reconstruct
    transaction and balance URLs from captured traffic without
    bank-specific shape assumptions.
  • IDashboardClickState / buildDashboardClickState — the click-at
    timestamp slot shared between DASHBOARD.ACTION (writer) and
    SCRAPE.PRE (reader).
  • ICollectionState / buildCollectionState — the live recording-gate
    accessors consulted on every captured response so the page
    listener can short-circuit during pre-auth phases without
    being torn down.

NetworkDiscovery.ts imports what it still needs (buildBalUrlFromTraffic,
buildCollectionState, buildDashboardClickState, buildTxnUrlFromTraffic,
IDashboardClickState) from EndpointState/EndpointState.js. Public
surface unchanged — these are all internal to the Network module.

Phase 4 split: commit 5/9. NetworkDiscovery.ts: 723 → 563 LoC.
